# MedWayHealthCare2.0/src/Presentation/ProductAdminPortal/Documentation/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: ProductAdminPortal API
  description: |
    Domain-agnostic product configuration and subscription management system.
    
    ## Features
    - Multi-tenant product catalog management
    - Flexible pricing engine with tiered, per-unit, and complexity-based pricing
    - Usage tracking and billing calculation
    - Subscription lifecycle management
    - Audit trails and versioning
    - RBAC with granular permissions
    
    ## Authentication
    All endpoints require JWT bearer token authentication.
    
    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for anonymous users
    
  version: 1.0.0
  contact:
    name: ProductAdminPortal Support
    email: support@productadminportal.com
servers:
  - url: https://api.productadminportal.com/v1
    description: Production server
  - url: https://staging-api.productadminportal.com/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
        billing_period_start:
          type: string
          format: date
        billing_period_end:
          type: string
          format: date
        subtotal:
          type: number
        tax_amount:
          type: number
        discount_amount:
          type: number
        total_amount:
          type: number
        currency_code:
          type: string
        status:
          type: string
          enum: [draft, pending, paid, overdue, cancelled]
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
    InvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        item_type:
          type: string
          enum: [base_fee, usage, one_time, discount, tax, adjustment]
        quantity:
          type: integer
        unit_price:
          type: number
        total_price:
          type: number
paths:
...
  /pricing/calculate:
    post:
      tags:
        - Pricing
      summary: Calculate pricing for usage
      description: |
        Calculate the invoice amount for a customer subscription based on usage records.
        This is the core pricing engine endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenant_id:
                  type: string
                  format: uuid
                customer_subscription_id:
                  type: string
                  format: uuid
                billing_period_start:
                  type: string
                  format: date
                billing_period_end:
                  type: string
                  format: date
                usage_records:
                  type: array
                  items:
                    type: object
                    properties:
                      entity_id:
                        type: string
                        format: uuid
                      units:
                        type: integer
                      complexity:
                        type: string
                        enum: [low, medium, high, critical]
                include_details:
                  type: boolean
                  default: true
              required:
                - tenant_id
                - customer_subscription_id
                - billing_period_start
                - billing_period_end
            example:
              tenant_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
              customer_subscription_id: "550e8400-e29b-41d4-a716-446655440000"
              billing_period_start: "2025-11-01"
              billing_period_end: "2025-11-30"
              usage_records: [
                {
                  entity_id: "entity-001",
                  units: 8000,
                  complexity: "low"
                },
                {
                  entity_id: "entity-001",
                  units: 3000,
                  complexity: "medium"
                },
                {
                  entity_id: "entity-001",
                  units: 1500,
                  complexity: "high"
                }
              ]
              include_details: true
      responses:
        '200':
          description: Pricing calculated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription_plan:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      base_fee:
                        type: number
                  line_items:
                    type: array
                    items:
                      type: object
                      properties:
                        description:
                          type: string
                        entity_id:
                          type: string
                        quantity:
                          type: integer
                        unit_price:
                          type: number
                        total_price:
                          type: number
                        calculation_details:
                          type: object
                  subtotal:
                    type: number
                  tax_amount:
                    type: number
                  discount_amount:
                    type: number
                  total_amount:
                    type: number
                  currency_code:
                    type: string
              example:
                subscription_plan:
                  id: "plan-gold-001"
                  name: "Gold Plan"
                  base_fee: 1000.00
                line_items:
                  - description: "Gold Plan - Base Fee"
                    entity_id: null
                    quantity: 1
                    unit_price: 1000.00
                    total_price: 1000.00
                    calculation_details:
                      type: "base_fee"
                  - description: "Patient Registration - Low Complexity (Tier 1: 1001-10000)"
                    entity_id: "entity-001"
                    quantity: 7000
                    unit_price: 0.10
                    total_price: 700.00
                    calculation_details:
                      tier: "1001-10000"
                      complexity: "low"
                      multiplier: 1.0
                  - description: "Patient Registration - Medium Complexity (Tier 1: 1001-10000)"
                    entity_id: "entity-001"
                    quantity: 2000
                    unit_price: 0.20
                    total_price: 400.00
                    calculation_details:
                      tier: "1001-10000"
                      complexity: "medium"
                      multiplier: 2.0
                  - description: "Patient Registration - Medium Complexity (Tier 2: >10000)"
                    entity_id: "entity-001"
                    quantity: 1000
                    unit_price: 0.14
                    total_price: 140.00
                    calculation_details:
                      tier: ">10000"
                      complexity: "medium"
                      multiplier: 2.0
                  - description: "Patient Registration - High Complexity (Tier 2: >10000)"
                    entity_id: "entity-001"
                    quantity: 1500
                    unit_price: 0.28
                    total_price: 420.00
                    calculation_details:
                      tier: ">10000"
                      complexity: "high"
                      multiplier: 4.0
                subtotal: 2660.00
                tax_amount: 479.00
                discount_amount: 0.00
                total_amount: 3139.00
                currency_code: "INR"

  # ===========================
  # SIMULATION
  # ===========================
  /simulate:
    post:
      tags:
        - Simulation
      summary: Simulate pricing for projected usage
      description: |
        Simulate what an invoice would look like for a given subscription with projected usage.
        Useful for customers to preview costs before committing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription_plan_id:
                  type: string
                  format: uuid
                projected_usage:
                  type: array
                  items:
                    type: object
                    properties:
                      entity_id:
                        type: string
                        format: uuid
                      units:
                        type: integer
                      complexity_breakdown:
                        type: object
                        properties:
                          low:
                            type: integer
                          medium:
                            type: integer
                          high:
                            type: integer
                billing_cycle:
                  type: string
                  enum: [monthly, yearly]
                discount_code:
                  type: string
              required:
                - subscription_plan_id
                - projected_usage
            example:
              subscription_plan_id: "plan-gold-001"
              projected_usage:
                - entity_id: "entity-001"
                  units: 12500
                  complexity_breakdown:
                    low: 8000
                    medium: 3000
                    high: 1500
              billing_cycle: "monthly"
              discount_code: "LAUNCH20"
      responses:
        '200':
          description: Simulation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  simulation_id:
                    type: string
                  subscription_plan:
                    type: object
                  projected_invoice:
                    $ref: '#/components/schemas/Invoice'
                  savings_vs_other_plans:
                    type: array
                    items:
                      type: object

  # ===========================
  # VERSIONING & AUDIT
  # ===========================
...